refactor: organize initial app
