+++ apps/project-management-simulation/src/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import '../../common/index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from '../../common/reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
