+++ index.html
@@ -1,35 +1,135 @@
+---
+layout: default
+title: Project Examples
+schema_type: CollectionPage
+tags: [Examples, Visualization, KnowledgeManagement]
+---
 <!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web site created using @vitejs/plugin-react"
-    />
-    <link rel="apple-touch-icon" href="/logo192.png" />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link rel="manifest" href="/manifest.json" />
-    <title>React App</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
+<html>
+<head>
+<meta charset="utf-8" />
+<title>Project Examples</title>
+<style>
+body{font-family:Arial, sans-serif;margin:20px;}
+.filter{margin-bottom:1em;}
+.filter button{margin-right:0.5em;padding:0.3em 0.6em;}
+.example-card{border:1px solid #ccc;border-radius:4px;padding:1em;margin:0.5em;display:inline-block;vertical-align:top;width:250px;}
+.example-card h3{margin-top:0;margin-bottom:0.5em;font-size:1.1em;}
+.example-card img{width:100%;height:auto;margin-bottom:0.5em;}
+</style>
+</head>
+<body>
+<h1>Project Examples</h1>
+<p><a href="https://lawrencerowland.github.io/">Home</a></p>
+<div class="filter" id="filter"></div>
+<div id="examples"></div>
+<script>
+function parseCSV(text){
+    const lines=text.trim().split(/\r?\n/).filter(l=>l.trim());
+    const headers=parseLine(lines[0].replace(/^\uFEFF/,''));
+    const data=[];
+    for(let i=1;i<lines.length;i++){
+        const row=parseLine(lines[i]);
+        if(row.length===headers.length){
+            const obj={};
+            headers.forEach((h,idx)=>obj[h]=row[idx].replace(/^"|"$/g,''));
+            data.push(obj);
+        }
+    }
+    return data;
+    function parseLine(line){
+        const result=[];let cur='';let inQuotes=false;
+        for(let i=0;i<line.length;i++){
+            const ch=line[i];
+            if(ch==='"'){
+                if(inQuotes && line[i+1]==='"'){cur+='"';i++;}
+                else inQuotes=!inQuotes;
+            }else if(ch===','&&!inQuotes){result.push(cur);cur='';}
+            else{cur+=ch;}
+        }
+        result.push(cur);
+        return result;
+    }
+}
 
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
+function createFilters(tags){
+    const container=document.getElementById('filter');
+    const allBtn=document.createElement('button');
+    allBtn.textContent='All';
+    allBtn.dataset.tag='all';
+    container.appendChild(allBtn);
+    tags.forEach(tag=>{
+        const btn=document.createElement('button');
+        btn.textContent=tag;
+        btn.dataset.tag=tag;
+        container.appendChild(btn);
+    });
+}
 
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-  </body>
-  <script type="module" src="/src/index.jsx"></script>
+function createCards(data){
+    const container=document.getElementById('examples');
+    data.forEach(item=>{
+        if(!item.name) return;
+        const card=document.createElement('div');
+        card.className='example-card';
+        card.dataset.tags=item.tags;
+        const img=document.createElement('img');
+        img.src='pics/'+item['#']+'.PNG';
+        img.alt=item.name;
+        img.onerror=()=>{img.src='pics/blank.png';};
+        card.appendChild(img);
+        const title=document.createElement('h3');
+        const link=document.createElement('a');
+        link.href='apps/'+item.name+'/index.html';
+        link.textContent=item.name;
+        title.appendChild(link);
+        card.appendChild(title);
+        const desc=document.createElement('p');
+        desc.textContent=item.description;
+        card.appendChild(desc);
+        const origin=document.createElement('p');
+        origin.innerHTML='<strong>Origin:</strong> '+item.origin;
+        card.appendChild(origin);
+        if(item.model){
+            const model=document.createElement('p');
+            model.innerHTML='<strong>Model:</strong> '+item.model;
+            card.appendChild(model);
+        }
+        const tags=document.createElement('p');
+        tags.innerHTML='<strong>Tags:</strong> '+item.tags;
+        card.appendChild(tags);
+        if(item['explainer post']){
+            const exp=document.createElement('p');
+            const l=document.createElement('a');
+            l.href=item['explainer post'];
+            l.textContent='Explainer post';
+            exp.appendChild(l);
+            card.appendChild(exp);
+        }
+        container.appendChild(card);
+    });
+}
+
+function filterExamples(tag){
+    document.querySelectorAll('.example-card').forEach(card=>{
+        const tags=card.dataset.tags.split(',').map(t=>t.trim());
+        if(tag==='all'||tags.includes(tag)) card.style.display='inline-block';
+        else card.style.display='none';
+    });
+}
+
+fetch('app-index.csv?t='+Date.now())
+  .then(resp=>resp.text())
+  .then(text=>{
+      const data=parseCSV(text);
+      const tags=[...new Set(data.flatMap(d=>d.tags.split(',').map(t=>t.trim())))]
+            .filter(t=>t);
+      createFilters(tags);
+      createCards(data);
+      document.querySelectorAll('.filter button').forEach(btn=>{
+          btn.addEventListener('click',()=>filterExamples(btn.dataset.tag));
+      });
+  });
+</script>
+</body>
 </html>
