+++ vite.config.js
@@ -1,16 +1,21 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import { dirname, resolve } from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  root: resolve(__dirname, 'apps/project-management-simulation'),
   base: process.env.NODE_ENV === 'production' ? '/React_proj-apps/' : '/',
   build: {
-    outDir: 'docs',
+    outDir: resolve(__dirname, 'docs/apps/project-management-simulation'),
   },
   plugins: [react()],
   test: {
     globals: true,
-    setupFiles: "./src/setupTests.js",
+    setupFiles: resolve(__dirname, 'src/common/setupTests.js'),
     environment: 'jsdom',
   },
 })
